@@ -71,17 +71,16 @@ class LineBodyState extends State<LineBody> {
   void _add() {
     controller.addLine(
       LineOptions(
-        geometry: [
-          LatLng(-33.86711, 151.1947171),
-          LatLng(-33.86711, 151.1947171),
-          LatLng(-32.86711, 151.1947171),
-          LatLng(-33.86711, 152.1947171),
-        ],
-        lineColor: "#ff0000",
-        lineWidth: 14.0,
-        lineOpacity: 0.5,
-        draggable: true
-      ),
+          geometry: [
+            LatLng(-33.86711, 151.1947171),
+            LatLng(-33.86711, 151.1947171),
+            LatLng(-32.86711, 151.1947171),
+            LatLng(-33.86711, 152.1947171),
+          ],
+          lineColor: "#ff0000",
+          lineWidth: 14.0,
+          lineOpacity: 0.5,
+          draggable: true),
     );
     setState(() {
       _lineCount += 1;
@@ -95,7 +94,6 @@ class LineBodyState extends State<LineBody> {
       _lineCount -= 1;
     });
   }
-
 
   Future<void> _changeAlpha() async {
     double current = _selectedLine.options.lineOpacity;
@@ -123,10 +121,7 @@ class LineBodyState extends State<LineBody> {
   void onStyleLoadedCallback() {
     controller.addLine(
       LineOptions(
-        geometry: [
-          LatLng(37.4220, -122.0841),
-          LatLng(37.4240, -122.0941)
-        ],
+        geometry: [LatLng(37.4220, -122.0841), LatLng(37.4240, -122.0941)],
         lineColor: "#ff0000",
         lineWidth: 14.0,
         lineOpacity: 0.5,
@@ -164,11 +159,11 @@ class LineBodyState extends State<LineBody> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
                           child: const Text('add'),
                           onPressed: (_lineCount == 12) ? null : _add,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('remove'),
                           onPressed: (_selectedLine == null) ? null : _remove,
                         ),
@@ -176,25 +171,27 @@ class LineBodyState extends State<LineBody> {
                     ),
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
                           child: const Text('change alpha'),
                           onPressed:
                               (_selectedLine == null) ? null : _changeAlpha,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('toggle visible'),
                           onPressed:
                               (_selectedLine == null) ? null : _toggleVisible,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('print current LatLng'),
-                          onPressed:
-                              (_selectedLine == null) ? null : () async{
-                                var latLngs = await controller.getLineLatLngs(_selectedLine);
-                                for (var latLng in latLngs) {
-                                  print(latLng.toString());
-                                }
-                              },
+                          onPressed: (_selectedLine == null)
+                              ? null
+                              : () async {
+                                  var latLngs = await controller
+                                      .getLineLatLngs(_selectedLine);
+                                  for (var latLng in latLngs) {
+                                    print(latLng.toString());
+                                  }
+                                },
                         ),
                       ],
                     ),
