@@ -120,7 +120,7 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
 
   void _getLatLng() async {
     LatLng latLng = await controller.getCircleLatLng(_selectedCircle);
-    Scaffold.of(context).showSnackBar(
+    ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
         content: Text(latLng.toString()),
       ),
@@ -145,7 +145,8 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
       // default value
       current = 0;
     }
-    _updateSelectedCircle(CircleOptions(circleStrokeWidth: current == 0 ? 5.0 : 0));
+    _updateSelectedCircle(
+        CircleOptions(circleStrokeWidth: current == 0 ? 5.0 : 0));
   }
 
   Future<void> _changeCircleStrokeColor() async {
@@ -156,7 +157,8 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     }
 
     _updateSelectedCircle(
-      CircleOptions(circleStrokeColor: current == "#FFFFFF" ? "#FF0000" : "#FFFFFF"),
+      CircleOptions(
+          circleStrokeColor: current == "#FFFFFF" ? "#FF0000" : "#FFFFFF"),
     );
   }
 
@@ -191,8 +193,7 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
     }
 
     _updateSelectedCircle(
-      CircleOptions(
-          circleColor: "#FFFF00"),
+      CircleOptions(circleColor: "#FFFF00"),
     );
   }
 
@@ -236,11 +237,11 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
                           child: const Text('add'),
                           onPressed: (_circleCount == 12) ? null : _add,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('remove'),
                           onPressed: (_selectedCircle == null) ? null : _remove,
                         ),
@@ -248,61 +249,64 @@ class PlaceCircleBodyState extends State<PlaceCircleBody> {
                     ),
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-opacity'),
-                          onPressed:
-                              (_selectedCircle == null) ? null : _changeCircleOpacity,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleOpacity,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-radius'),
                           onPressed: (_selectedCircle == null)
                               ? null
                               : _changeCircleRadius,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-color'),
-                          onPressed:
-                          (_selectedCircle == null) ? null : _changeCircleColor,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleColor,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-blur'),
-                          onPressed:
-                          (_selectedCircle == null) ? null : _changeCircleBlur,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleBlur,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-stroke-width'),
-                          onPressed:
-                              (_selectedCircle == null) ? null : _changeCircleStrokeWidth,
+                          onPressed: (_selectedCircle == null)
+                              ? null
+                              : _changeCircleStrokeWidth,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-stroke-color'),
                           onPressed: (_selectedCircle == null)
                               ? null
                               : _changeCircleStrokeColor,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change circle-stroke-opacity'),
                           onPressed: (_selectedCircle == null)
                               ? null
                               : _changeCircleStrokeOpacity,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change position'),
                           onPressed: (_selectedCircle == null)
                               ? null
                               : _changePosition,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('toggle draggable'),
                           onPressed: (_selectedCircle == null)
                               ? null
                               : _changeDraggable,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('get current LatLng'),
-                          onPressed: (_selectedCircle == null)
-                              ? null
-                              : _getLatLng,
+                          onPressed:
+                              (_selectedCircle == null) ? null : _getLatLng,
                         ),
                       ],
                     ),
