@@ -5,8 +5,7 @@ import 'main.dart';
 import 'page.dart';
 
 class FullMapPage extends ExamplePage {
-  FullMapPage()
-      : super(const Icon(Icons.map), 'Full screen map');
+  FullMapPage() : super(const Icon(Icons.map), 'Full screen map');
 
   @override
   Widget build(BuildContext context) {
@@ -32,11 +31,12 @@ class FullMapState extends State<FullMap> {
   Widget build(BuildContext context) {
     return new Scaffold(
         body: MapboxMap(
-          accessToken: MapsDemo.ACCESS_TOKEN,
-          onMapCreated: _onMapCreated,
-          initialCameraPosition:
-          const CameraPosition(target: LatLng(0.0, 0.0)),
-        )
-    );
+      accessToken: MapsDemo.ACCESS_TOKEN,
+      onMapCreated: _onMapCreated,
+      initialCameraPosition: const CameraPosition(target: LatLng(0.0, 0.0)),
+      onStyleLoadedCallback: onStyleLoadedCallback,
+    ));
   }
+
+  void onStyleLoadedCallback() {}
 }
