@@ -2,15 +2,14 @@
 // Generated file. Do not edit.
 //
 
-// ignore: unused_import
-import 'dart:ui';
+// ignore_for_file: lines_longer_than_80_chars
 
 import 'package:mapbox_gl_web/mapbox_gl_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 // ignore: public_member_api_docs
-void registerPlugins(PluginRegistry registry) {
-  MapboxMapPlugin.registerWith(registry.registrarFor(MapboxMapPlugin));
-  registry.registerMessageHandler();
+void registerPlugins(Registrar registrar) {
+  MapboxMapPlugin.registerWith(registrar);
+  registrar.registerMessageHandler();
 }
