@@ -5,11 +5,16 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:location/location.dart';
+import 'package:mapbox_gl_example/custom_marker.dart';
 import 'package:mapbox_gl_example/full_map.dart';
+import 'package:mapbox_gl_example/offline_regions.dart';
+import 'package:mapbox_gl_example/place_batch.dart';
 
 import 'animate_camera.dart';
+import 'annotation_order_maps.dart';
 import 'full_map.dart';
 import 'line.dart';
+import 'local_style.dart';
 import 'map_ui.dart';
 import 'move_camera.dart';
 import 'page.dart';
@@ -27,13 +32,17 @@ final List<ExamplePage> _allPages = <ExamplePage>[
   PlaceSymbolPage(),
   PlaceSourcePage(),
   LinePage(),
+  LocalStylePage(),
   PlaceCirclePage(),
   PlaceFillPage(),
   ScrollingMapPage(),
+  OfflineRegionsPage(),
+  AnnotationOrderPage(),
+  CustomMarkerPage(),
+  BatchAddPage(),
 ];
 
 class MapsDemo extends StatelessWidget {
-
   //FIXME: Add your Mapbox access token here
   static const String ACCESS_TOKEN = "YOUR_TOKEN_HERE";
 
